requirements.txt
pandas
numpy
scikit-learn
javalang
pyyaml
tqdm
google-genai
python-dotenv
